feat : user 커밋 (#8)
* feat : 회원탈퇴 API

* fix : UserInfo userIdx 타입 Long->Integer 전체 수정

* feat : 유저 정보 조회 API

* feat : 유저 정보 수정 API

* fix : PostSignUpReq 파라미터명 parameters -> postSignUpReq 변경

* fix : Validation 명칭 변경

+++ src/main/java/com/example/ged/src/user/models/PostUserRes.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class PostUserRes {
-    private final Long userIdx;
+    private final Integer userIdx;
     private final String jwt;
 }
